Move generateRedirectUri into separate file in /authorise dir
As it is a generic function reused by both accounts and payments requests.
