+++ app/authorise/index.js
@@ -1,4 +1,5 @@
 const { createClaims, createJsonWebSignature } = require('./request-jws');
+const { generateRedirectUri } = require('./authorise-uri');
 const { setTokenPayload, accessToken } = require('./access-tokens');
 const { authorisationCodeGrantedHandler } = require('./authorisation-code-granted');
 
@@ -6,4 +7,5 @@ exports.accessToken = accessToken;
 exports.authorisationCodeGrantedHandler = authorisationCodeGrantedHandler;
 exports.createClaims = createClaims;
 exports.createJsonWebSignature = createJsonWebSignature;
+exports.generateRedirectUri = generateRedirectUri;
 exports.setTokenPayload = setTokenPayload;
