+++ app/setup-account-request/account-request-authorise-consent.js
@@ -1,42 +1,7 @@
 const { setupAccountRequest } = require('./setup-account-request');
-const { createClaims, createJsonWebSignature } = require('../authorise');
-const { authorisationEndpoint, getClientCredentials, issuer } = require('../authorisation-servers');
+const { generateRedirectUri } = require('../authorise');
 const error = require('debug')('error');
 const debug = require('debug')('debug');
-const env = require('env-var');
-const qs = require('qs');
-
-const registeredRedirectUrl = env.get('SOFTWARE_STATEMENT_REDIRECT_URL').asString();
-
-const statePayload = (authorisationServerId, sessionId) => {
-  const state = {
-    authorisationServerId,
-    sessionId,
-  };
-  return Buffer.from(JSON.stringify(state)).toString('base64');
-};
-
-const generateRedirectUri = async (authorisationServerId, requestId, scope, sessionId) => {
-  const { clientId } = await getClientCredentials(authorisationServerId);
-  const state = statePayload(authorisationServerId, sessionId);
-  const authEndpoint = await authorisationEndpoint(authorisationServerId);
-  const authServerIssuer = await issuer(authorisationServerId);
-  const payload = createClaims(
-    scope, requestId, clientId, authServerIssuer,
-    registeredRedirectUrl, state, createClaims,
-  );
-  const signature = createJsonWebSignature(payload);
-  const uri =
-    `${authEndpoint}?${qs.stringify({
-      redirect_uri: registeredRedirectUrl,
-      state,
-      client_id: clientId,
-      response_type: 'code',
-      request: signature,
-      scope,
-    })}`;
-  return uri;
-};
 
 const accountRequestAuthoriseConsent = async (req, res) => {
   res.setHeader('Access-Control-Allow-Origin', '*');
@@ -58,6 +23,4 @@ const accountRequestAuthoriseConsent = async (req, res) => {
   }
 };
 
-exports.statePayload = statePayload;
 exports.accountRequestAuthoriseConsent = accountRequestAuthoriseConsent;
-exports.generateRedirectUri = generateRedirectUri;
